Move dataset util funcs into separate module
refactor imports and tests accordingly
